+++ src/lib/router.ts
@@ -1,6 +1,12 @@
 export const shouldShowNavigation = (pathname: string): boolean => {
   // Navigation을 숨길 경로 패턴들
-  const hidePatterns = ['/onboarding', '/login', '/user/auth'];
+  const hidePatterns = [
+    '/onboarding',
+    '/login',
+    '/user/auth',
+    '/signup/email',
+    '/signup/profile',
+  ];
 
   return !hidePatterns.some(
     (pattern) => pathname === pattern || pathname.startsWith(`${pattern}/`),
